move(common): common 패키지 분리
